- type definition fixed - autoconf build variable $(top_builddir) fix

+++ lib/buffer/buffer_default.c
@@ -1,7 +1,7 @@
 #include "buffer.h"
 
 /* set sane defaults + operation */
-void buffer_default(buffer *b, int (*op)())
+void buffer_default(buffer *b, ssize_t (*op)())
 {
   b->fd = -1;
   b->p = 0;
